extension: use helper to connect signals
Use the custom `_connect` helper to connect the signals for
`state-changed` and `updated` signals, so that they get
properly disconnected when the object gets destroyed. This
should fix the issue where signals were delivered after the
extension was destroyed and the signal hanlder was trying
to set properties of the (already disposed) actor leading
to a stack-trace.